+++ src/Indice.Features.Cases.App/src/app/features/case-types/case-type-create/case-type-create.component.ts
@@ -9,16 +9,10 @@ import { CaseTypesService } from '../case-types.service';
 })
 export class CaseTypeCreateComponent implements OnInit {
 
-  
-
   public widgets = this.caseTypesService.widgets;
-
   public framework = this.caseTypesService.framework;
-
   public schema = this.caseTypesService.schema;
-
   public layout = this.caseTypesService.onLoadLayout();
-
   public data: any = {};
 
   constructor(private caseTypesService: CaseTypesService) { }