+++ src/Indice.Features.Cases.AspNetCore/Controllers/MyCasesController.cs
@@ -1,5 +1,4 @@
 ﻿using System.Net.Mime;
-using System.Text.Json;
 using Indice.AspNetCore.Filters;
 using Indice.Features.Cases.Events;
 using Indice.Features.Cases.Interfaces;
@@ -150,19 +149,9 @@ namespace Indice.Features.Cases.Controllers
         }
 
         private async Task<byte[]> CreatePdf(Case @case) {
-            var isPortrait = true;
-            var digitallySigned = false;
-            if (@case.CaseType.Config is not null) {
-                var caseTypeConfig = JsonSerializer.Deserialize<JsonDocument>(@case.CaseType.Config);
-                if (caseTypeConfig.RootElement.TryGetProperty("IsPortrait", out var isPortraitConfig)) {
-                    isPortrait = isPortraitConfig.GetBoolean();
-                }
-                if (caseTypeConfig.RootElement.TryGetProperty("DigitallySigned", out var digitallySignedConfig)) {
-                    digitallySigned = digitallySignedConfig.GetBoolean();
-                }
-            }
             var template = await _caseTemplateService.RenderTemplateAsync(@case);
-            return await _casePdfService.HtmlToPdfAsync(template, isPortrait, digitallySigned);
+            var pdfOptions = new PdfOptions(@case.CaseType.Config);
+            return await _casePdfService.HtmlToPdfAsync(template, pdfOptions, @case);
         }
     }
 }