+++ src/Indice.Features.Cases.AspNetCore/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##  [6.10.4] - 2023-03-02
+### Changed
+- Case type form for create/update is simplified. Only case type entity is being edited.
+
+## [6.10.3] - 2023-03-01
+### Added
+- `PdfOptions` as a case type Config option
+
+## [6.10.1] - 2023-02-15
+### Added
+- Reports controller to handle UI dashboard stats
+
 ## [6.9.1] - 2023-02-07
 ### Changed
 - Notification Subscriptions are now per case type